Trait chain_crypto::bech32::Bech32

pub trait Bech32 {
    const BECH32_HRP: &'static str;
    const BYTES_LEN: usize;

    // Required methods
    fn try_from_bech32_str(bech32_str: &str) -> Result<Self>
       where Self: Sized;
    fn to_bech32_str(&self) -> String;
}

Bech32 encoding for fixed-size binary objects.

Required Associated Constants

const BECH32_HRP: &'static str

The human-readable prefix that is used to represent the the object in the Bech32 format. On decoding, the HRP of the input string is checked against this value.

const BYTES_LEN: usize

Length of the encoded binary data.

Required Methods

fn try_from_bech32_str(bech32_str: &str) -> Result<Self>where Self: Sized,

Decodes the object from its Bech32 string representation.

fn to_bech32_str(&self) -> String

Produces a Bech32 string format representation of the object.